requirements.txt
pymongo==4.6.1
pytest==7.4.0
pytest-mock==3.11.1
pytest-cov==4.1.0
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
langchain==0.0.335
pyyaml==6.0.1
langchain-groq==0.0.1
openai==1.3.5
pillow==10.1.0
langchain-openrouter==0.0.1
black==23.7.0
isort==5.12.0
flake8==6.1.0
httpx==0.24.1 